refactor: remove old register files and update index files
